wechat/requirements.txt
requests
PyYAML
Pillow
watchdog
Markdown
premailer
beautifulsoup4
python-dotenv
lxml
